move more fns to sc_light
