+++ client/light/src/lib.rs
@@ -18,9 +18,40 @@
 
 //! Light client components.
 
+use sp_runtime::traits::{Block as BlockT, HashFor};
+use sc_client_api::CloneableSpawn;
+use std::sync::Arc;
+use sp_core::traits::CodeExecutor;
+
 pub mod backend;
 pub mod blockchain;
 pub mod call_executor;
 pub mod fetcher;
 
 pub use {backend::*, blockchain::*, call_executor::*, fetcher::*};
+
+/// Create an instance of fetch data checker.
+pub fn new_fetch_checker<E, B: BlockT, S: BlockchainStorage<B>>(
+	blockchain: Arc<Blockchain<S>>,
+	executor: E,
+	spawn_handle: Box<dyn CloneableSpawn>,
+) -> LightDataChecker<E, HashFor<B>, B, S>
+	where
+		E: CodeExecutor,
+{
+	LightDataChecker::new(blockchain, executor, spawn_handle)
+}
+
+/// Create an instance of light client blockchain backend.
+pub fn new_light_blockchain<B: BlockT, S: BlockchainStorage<B>>(storage: S) -> Arc<Blockchain<S>> {
+	Arc::new(Blockchain::new(storage))
+}
+
+/// Create an instance of light client backend.
+pub fn new_light_backend<B, S>(blockchain: Arc<Blockchain<S>>) -> Arc<Backend<S, HashFor<B>>>
+	where
+		B: BlockT,
+		S: BlockchainStorage<B>,
+{
+	Arc::new(Backend::new(blockchain))
+}
